Create a library of gadgets (#129)
* Create gadgets lib

* Cargo fix

* Create gadgets inside sudoku

* Cargo fix

* Separate between simple sudoku and taiga sudoku

* Fix simple_sudoku

* Cargo fix

* Fix clippy error

* Update rust toolchain
