+++ rust-toolchain
@@ -1,1 +1,1 @@
-1.58
+1.67.1
