+++ taiga_halo2/examples/simple_sudoku/main.rs
@@ -1,2 +1,4 @@
 pub mod circuit;
 pub mod vp;
+
+pub fn main() {}
