--- taiga_halo2/examples/sudoku/app/mod.rs
@@ -1,5 +1,0 @@
-pub mod dealer_intent_vp;
-pub mod intent_app_vp;
-pub mod sudoku_app_vp;
-pub mod valid_puzzle;
-pub mod valid_sudoku;
