--- taiga_halo2/examples/sudoku/app/valid_puzzle/vp.rs
@@ -1,1 +1,0 @@
-
