--- taiga_halo2/examples/sudoku/app/valid_sudoku/mod.rs
@@ -1,2 +1,0 @@
-pub mod circuit;
-pub mod vp;
