+++ taiga_halo2/examples/taiga_sudoku/main.rs
@@ -1,0 +1,8 @@
+pub mod app_vp;
+pub mod dealer_intent_vp;
+pub mod gadgets;
+pub mod intent_app_vp;
+use crate::app_vp::halo2_sudoku_app_vp_circuit_final;
+fn main() {
+    halo2_sudoku_app_vp_circuit_final();
+}
