+++ taiga_halo2/src/circuit/gadgets/add.rs
@@ -1,0 +1,111 @@
+use halo2_proofs::{
+    circuit::{AssignedCell, Chip, Layouter, Region},
+    plonk::{Advice, Column, ConstraintSystem, Error, Selector},
+    poly::Rotation,
+};
+use pasta_curves::arithmetic::FieldExt;
+
+use std::marker::PhantomData;
+
+// AddChip copy from halo2 example two-chip
+#[derive(Clone, Debug)]
+pub struct AddChip<F: FieldExt> {
+    config: AddConfig,
+    _marker: PhantomData<F>,
+}
+
+#[derive(Clone, Debug)]
+pub struct AddConfig {
+    advice: [Column<Advice>; 2],
+    s_add: Selector,
+}
+
+impl<F: FieldExt> Chip<F> for AddChip<F> {
+    type Config = AddConfig;
+    type Loaded = ();
+
+    fn config(&self) -> &Self::Config {
+        &self.config
+    }
+
+    fn loaded(&self) -> &Self::Loaded {
+        &()
+    }
+}
+
+impl<F: FieldExt> AddChip<F> {
+    pub fn construct(
+        config: <Self as Chip<F>>::Config,
+        _loaded: <Self as Chip<F>>::Loaded,
+    ) -> Self {
+        Self {
+            config,
+            _marker: PhantomData,
+        }
+    }
+
+    pub fn configure(
+        meta: &mut ConstraintSystem<F>,
+        advice: [Column<Advice>; 2],
+    ) -> <Self as Chip<F>>::Config {
+        let s_add = meta.selector();
+
+        // Define our addition gate!
+        meta.create_gate("add", |meta| {
+            let lhs = meta.query_advice(advice[0], Rotation::cur());
+            let rhs = meta.query_advice(advice[1], Rotation::cur());
+            let out = meta.query_advice(advice[0], Rotation::next());
+            let s_add = meta.query_selector(s_add);
+
+            vec![s_add * (lhs + rhs - out)]
+        });
+
+        AddConfig { advice, s_add }
+    }
+}
+
+pub trait AddInstructions<F: FieldExt>: Chip<F> {
+    /// Returns `c = a + b`.
+    fn add(
+        &self,
+        layouter: impl Layouter<F>,
+        a: &AssignedCell<F, F>,
+        b: &AssignedCell<F, F>,
+    ) -> Result<AssignedCell<F, F>, Error>;
+}
+
+impl<F: FieldExt> AddInstructions<F> for AddChip<F> {
+    fn add(
+        &self,
+        mut layouter: impl Layouter<F>,
+        a: &AssignedCell<F, F>,
+        b: &AssignedCell<F, F>,
+    ) -> Result<AssignedCell<F, F>, Error> {
+        let config = self.config();
+
+        layouter.assign_region(
+            || "add",
+            |mut region: Region<'_, F>| {
+                // We only want to use a single addition gate in this region,
+                // so we enable it at region offset 0; this means it will constrain
+                // cells at offsets 0 and 1.
+                config.s_add.enable(&mut region, 0)?;
+
+                // The inputs we've been given could be located anywhere in the circuit,
+                // but we can only rely on relative offsets inside this region. So we
+                // assign new cells inside the region and constrain them to have the
+                // same values as the inputs.
+                a.copy_advice(|| "lhs", &mut region, config.advice[0], 0)?;
+                b.copy_advice(|| "rhs", &mut region, config.advice[1], 0)?;
+
+                // Now we can compute the addition result, which is to be assigned
+                // into the output position.
+                let value = a.value().copied() + b.value();
+
+                // Finally, we do the assignment to the output, returning a
+                // variable to be used in another part of the circuit.
+                region.assign_advice(|| "lhs + rhs", config.advice[0], 1, || value)
+            },
+        )
+    }
+}
