+++ taiga_halo2/src/circuit/gadgets/mul.rs
@@ -1,0 +1,101 @@
+use halo2_proofs::{
+    circuit::{AssignedCell, Chip, Layouter, Region},
+    plonk::{Advice, Column, ConstraintSystem, Error, Selector},
+    poly::Rotation,
+};
+use pasta_curves::arithmetic::FieldExt;
+
+use std::marker::PhantomData;
+
+/// An instruction set for multiplying two circuit words (field elements).
+pub trait MulInstructions<F: FieldExt>: Chip<F> {
+    /// Constraints `a * b` and returns the multiplication.
+    fn mul(
+        &self,
+        layouter: impl Layouter<F>,
+        a: &AssignedCell<F, F>,
+        b: &AssignedCell<F, F>,
+    ) -> Result<AssignedCell<F, F>, Error>;
+}
+
+#[derive(Clone, Debug)]
+pub struct MulConfig {
+    advice: [Column<Advice>; 2],
+    s_mul: Selector,
+}
+
+/// A chip implementing a single multiplication constraint `c = a * b` on a single row.
+pub struct MulChip<F: FieldExt> {
+    config: MulConfig,
+    _marker: PhantomData<F>,
+}
+
+impl<F: FieldExt> Chip<F> for MulChip<F> {
+    type Config = MulConfig;
+    type Loaded = ();
+
+    fn config(&self) -> &Self::Config {
+        &self.config
+    }
+
+    fn loaded(&self) -> &Self::Loaded {
+        &()
+    }
+}
+
+impl<F: FieldExt> MulChip<F> {
+    pub fn configure(meta: &mut ConstraintSystem<F>, advice: [Column<Advice>; 2]) -> MulConfig {
+        let s_mul = meta.selector();
+        meta.create_gate("Field element multiplication: c = a * b", |meta| {
+            let s_mul = meta.query_selector(s_mul);
+            let lhs = meta.query_advice(advice[0], Rotation::cur());
+            let rhs = meta.query_advice(advice[1], Rotation::cur());
+            let out = meta.query_advice(advice[0], Rotation::next());
+
+            vec![s_mul * (lhs * rhs - out)]
+        });
+
+        MulConfig { advice, s_mul }
+    }
+
+    pub fn construct(config: MulConfig) -> Self {
+        Self {
+            config,
+            _marker: PhantomData,
+        }
+    }
+}
+
+impl<F: FieldExt> MulInstructions<F> for MulChip<F> {
+    fn mul(
+        &self,
+        mut layouter: impl Layouter<F>,
+        a: &AssignedCell<F, F>,
+        b: &AssignedCell<F, F>,
+    ) -> Result<AssignedCell<F, F>, Error> {
+        layouter.assign_region(
+            || "mul",
+            |mut region: Region<'_, F>| {
+                // We only want to use a single addition gate in this region,
+                // so we enable it at region offset 0; this means it will constrain
+                // cells at offsets 0 and 1.
+                self.config.s_mul.enable(&mut region, 0)?;
+
+                // The inputs we've been given could be located anywhere in the circuit,
+                // but we can only rely on relative offsets inside this region. So we
+                // assign new cells inside the region and constrain them to have the
+                // same values as the inputs.
+                a.copy_advice(|| "lhs", &mut region, self.config.advice[0], 0)?;
+                b.copy_advice(|| "rhs", &mut region, self.config.advice[1], 0)?;
+
+                // Now we can compute the addition result, which is to be assigned
+                // into the output position.
+                let value = a.value().copied() * b.value();
+
+                // Finally, we do the assignment to the output, returning a
+                // variable to be used in another part of the circuit.
+                region.assign_advice(|| "lhs * rhs", self.config.advice[0], 1, || value)
+            },
+        )
+    }
+}
